+++ BugsnagPerformance/Assets/BugsnagPerformance/Scripts/Public/PerformanceConfiguration.cs
@@ -24,6 +24,7 @@ namespace BugsnagUnityPerformance
         public string AppVersion = Application.version;
         public int VersionCode = -1;
         public string BundleVersion;
+      
         public bool GenerateAnonymousId = true;
 
         public string[] EnabledReleaseStages;
@@ -34,10 +35,13 @@ namespace BugsnagUnityPerformance
 
         public string ReleaseStage = Debug.isDebugBuild ? "development" : "production";
 
+        public String[] TracePropagationUrlMatchPatterns;
+
         public PerformanceConfiguration(string apiKey)
         {
             ApiKey = apiKey;
         }
+
       
     }
 }