+++ BugsnagPerformance/Assets/BugsnagPerformance/Scripts/Public/Span.cs
@@ -22,6 +22,7 @@ namespace BugsnagUnityPerformance
         private object _endLock = new object();
         private OnSpanEnd _onSpanEnd;
         internal bool IsAppStartSpan;
+        internal bool WasAborted;
 
         public Span(string name, SpanKind kind, string id, string traceId, string parentSpanId, DateTimeOffset startTime, bool? isFirstClass, OnSpanEnd onSpanEnd)
         {
@@ -39,21 +40,7 @@ namespace BugsnagUnityPerformance
             _onSpanEnd = onSpanEnd;
         }
 
-        public void End()
-        {
-            lock (_endLock)
-            {
-                if (Ended)
-                {
-                    return;
-                }
-                Ended = true;
-            }
-            EndTime = DateTimeOffset.UtcNow;
-            _onSpanEnd(this);
-        }
-
-        internal void End(DateTimeOffset? endTime)
+        public void End(DateTimeOffset? endTime = null)
         {
             lock (_endLock)
             {
@@ -75,7 +62,9 @@ namespace BugsnagUnityPerformance
                 {
                     return;
                 }
+                WasAborted = true;
                 Ended = true;
+                _onSpanEnd(this);
             }
         }
 
@@ -107,7 +96,7 @@ namespace BugsnagUnityPerformance
             _onSpanEnd(this);
         }
 
-        public void EndNetworkSpan(int statusCode = -1, int requestContentLength = -1, int responseContentLength = -1)
+        public void EndNetworkSpan(int statusCode = -1, int requestContentLength = -1, int responseContentLength = -1, DateTimeOffset? endTime = null)
         {
             lock (_endLock)
             {
@@ -118,7 +107,7 @@ namespace BugsnagUnityPerformance
                 Ended = true;
             }
 
-            EndTime = DateTimeOffset.UtcNow;
+            EndTime = endTime == null ? DateTimeOffset.UtcNow : endTime.Value;
 
             if (statusCode > -1)
             {