+++ CHANGELOG.md
@@ -1,8 +1,23 @@
 # Changelog
 
-### TBD
+## TBD 
 
+### Additions
+
+- Added the trace parent header to requests made via the Bugsnag request wrapper and allow configuration via new TracePropagationUrls property. [#109](https://github.com/bugsnag/bugsnag-unity-performance/pull/109)
 - Added methods to generate anonymous id value [#92](https://github.com/bugsnag/bugsnag-unity-performance/pull/92)
+
+## v1.3.4 (2024-04-29)
+
+### Additions
+
+- Discard open spans when app is backgrounded [#105](https://github.com/bugsnag/bugsnag-unity-performance/pull/105)
+
+### Bug Fixes
+
+- Fixed issue where spans could not be ended with a custom end time. [#106](https://github.com/bugsnag/bugsnag-unity-performance/pull/106)
+
+- Fixed issue where custom spans with the SpanOption IsFirstClass set to false still had it reported as true. [#107](https://github.com/bugsnag/bugsnag-unity-performance/pull/107)
 
 ## v1.3.3 (2024-03-07)